v4l: CB2 mpeg2 ts abnormal & h265 seek freeze & mjpeg freeze. [1/1]
PD#SWPL-90247
PD#SWPL-89897
PD#SWPL-80917

Problem:
1.For a packet of one field of interlaced stream,
the decoder covers the vf information from
the second field over the first field,
resulting in only the ts information of the second field being output.
2.The resolution changes after h265 seek,
and the judgment condition is not perfect,
which leads to the failure to obtain the buffer.
3.MJPEG supports basic playback, seek,
fast forward and back, resolution switching.

Solution:
1.The vf information of each field of
the interlaced code stream is stored separately.
2.Improve the judgment conditions for obtaining the buffer,
so that the buffer can be obtained normally
after changing the resolution after seeking.
3.support mjpeg.

Verify:
sc2

Change-Id: I9ef0f7f3f603e540e752a6c3d54468997c3de455
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>