mpeg2: CB2 cts adaptive mpeg2 timeout & task leak. [1/1]
PD#SWPL-92708
PD#SWPL-93603

Problem:
1.When the resolution is changed,
the frame return of the upper layer is too slow,
and the decoder has been waiting
for the frame return in the interrupt function,
resulting in an exception of the decoder hardware timeout.
2.When the buffer is released, the task item is not released,
causing the task chain memory leak.

Solution:
1.While the interrupt thread is waiting to return the frame,
the timeout interrupt is not allowed to be processed.
2.Release the task item before
releasing the task chain memory.

Verify:
ohm

Change-Id: I8992b2de0dc2eab323d8a9f776bb0a2e54162c65
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>