v4l: CB2 Report the timestamp of the error frame. [1/1]
PD#SWPL-93731
PD#SWPL-96966
PD#SWPL-96709

Problem:
1.The error stream appears blurry.
2.Abnormal size code stream causes crash.
3.After the h265 error frame output policy is applied,
because the barcode stream has errors and
the reference frame is abnormal,
the dpb is full and the dpb slot is insufficient.

Solution:
1.Allows the internal reset operation of the error code stream,
and reports the timestamp of the error frame at the same time.
2.decoder supports streams with sizes larger than 64x64.
3.Do not output error frames,
and use the event of reporting error frames for processing.

Verify:
sc2

Change-Id: I7c41b0c5081f510bdfe16a9ddae614b2014490e8
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>