h265/avs: CB2 h265 error frame  & avs di issue & fg crash  [1/1]
PD#SWPL-101752
PD#SWPL-96742
PD#SWPL-101561
PD#SWPL-101776

Problem:
1.The h265 error data is in the same frame as the normal data,
but the decoder reports an error frame event,
resulting in no output of this frame.
2.The specific avs stream is jagged;
the stream syntax is progressive
but the frame is actually interlaced without di processing.
3.fg table is not initialized, causing it to crash.
4.avs ts output repeat

Solution:
1.Error frame events are not reported when error data
and normal data are in the same frame.
2.For this kind of code stream, forced interlacing
is the same as non-vl42.
3.initial fg table
4.Optimize ts management

Verify:
sc2

Change-Id: I8d014deb2f7f353884105f0a8346c6d82f8d5095
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>