vh265: CB2 H265 can't get eos buffer and freeze. [1/1]
PD#SWPL-100714

Problem:
When the seek reaches the end of the stream,
the stream is dirty data,
the decoder does not get the sequence header,
and the dpb initialization is not performed.
When the eos is sent,
the decoder gets the dpb buffer and causes an exception.

Solution:
For the case of sending eos without dpb initialization,
the decoder is also allowed to get dpb buffer.

Verify:
ah212

Change-Id: Ia9069819df7154039f63a04693a8df66ec64e1d3
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>