v4l: CB1 fixed issue of playback is not smooth. [1/1]
PD#SWPL-95844

Problem:
fixed issue of playback is not smooth

Solution:
The reason why multi-channel playback is not smooth
is that the cpu is busy and cannot be scheduled. Previously,
the task is not submitted immediately when data appears eagain.
Now, the cpu can be released for a short time so that
other processes can be scheduled in time.

Verify:
ohm

Change-Id: I1a01b3d5df8d7aa78b375bc8e6b730dc60f4b5fc
Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>