v4l: CF1 Optimized the scheduling of es buffer queue. [1/1]
PD#SWPL-106104

Problem:
After eagain is triggered, the es buffer queue will
be trapped in the kworker loop call, similar to while(1),
which consumes a large amount of cpu resources and
causes system performance problems.

Solution:
Change the mode of the kworker called in a loop
to wait for message notification, and the task of
scheduling es buffer will be started only
when the data really needs to be processed.

Verify:
ohm

Change-Id: I4757cd16de6e65ce0b8056f21603fa75f617afc0
Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>