decoder: CF2 fix the unsmooth problem for special streams. [1/1]
PD#SWPL-112719

Problem:
The input pts of the stream are continuous, causing the output pts
of the lookup to jump, resulting in unsmooth playback.

Solution:
Use bit62 and 63 of the pts_us64 variable in the vf structure to
transmit frame type information, and let the tsplayer process these pts
abnormal streams according to the frame type information.

Verify:
S5

Change-Id: I10b6c37234d91d6a1b0a9707c2bbd384cbce3703
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>