userdata: CB2 After pip is turned on, the main channel cc is lost [1/1]
PD#SWPL-110955

Problem:
The upper layer will refer to the pip channel as the player id,
causing the main channel player id to be overwritten,
and the decoder did not report the main channel cc.

Solution:
Adjust the decoder reporting cc strategy,
both the main channel and the pip channel cc are reported,
and the upper layer can choose the main channel cc.

Verify:
sc2

Change-Id: I599cb2b03928781d6f62baf02c1d5e90de50faef
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>