avs3: CB2 fix playback stuck problem due to packet loss. [1/1]
PD#SWPL-115612

Problem:
When packet loss occurs on the network, the decoding will lack the
reference frame.In this case, the reference relationship of the previous
frame will be updated, resulting in an abnormal backref.

Solution:
When an exception occurs, there is no need to update the reference
relationship of pic.
Porting 5.4 branch avs3 code to 5.15 branch.

Verify:
S5

Change-Id: I321ea396fb4087f6b715d027d11db69078220c16
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>