h265: CB2 h264 error stream freez. [1/1]
PD#SWPL-117580

Problem:
The stream is wrong, the DPB reference frame
cannot be swiped out, and the dpb slot is not enough,
resulting in a stuck.

Solution:
If the dpb slot is not enough and the number of reference frames
in the dpb is greater than the expected number, perform a buffer reset.

Verify:
ohm

Change-Id: I40beab30291f43b709a27039da3be836689c6cc1
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>