v4l: CB1 dw mmu box dynamic allocation. [1/1]
PD#SWPL-117487

Problem:
Since the dw mmu box is allocated at the start of the broadcast,
when the scene where the fg stream is switched to a non-fg stream,
it will not be able to play due to unallocated memory.

Solution:
Add dw mmu box dynamic allocation function.

Verify:
s4/s4d

Change-Id: Ia580c8a9e4f8ee18fde0d2c8897db82044c797f5
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>