mpeg4: CB1 fix some streams playback mosaic in stream mode. [1/1]
PD#SWPL-118155

Problem:
Since the bit 8 of the VLD_MEM_VIFIFO_CONTROL register is not 0, the
ucode reports the pic done interrupt here, but the parse module is
still working at this time. This will cause a decoding exception in the
subsequent decoding process.

Solution:
Ensure that the parse module is not busy by checking that bit 8 of the
VLD_MEM_VIFIFO_CONTROL register is not 1.

Verify:
S5

SourceCode:
ucode:
0.4.47-g4206e84
Id3122
I72ddb
I3367c
Ic48cf
I70f34

Change-Id: Ie020c33c0c4ff2a2207d03cf0f2c82cb5ca76e76
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>