mpeg12: CB1 fix the stuck playback problem of udp push special stream. [1/1]
PD#SWPL-74611

Problem:
Since a frame of data also includes information such as sequence end and
sequence header, the decoding of this frame of data enters the logic of
a packet of multiple frames. When decoding this frame of data again, due
to the logic of sequence end, the interrupt will be triggered again, so
that this frame of data will be decoded all the time and the playback
will be stuck.

Solution:
1.After searching to the start code, modify the inaccurate problem of
VIFF_BIT_CNT according to the alignment logic.
2.If the picture header is not parsed in the sequence end logic,
continue to search for the start code.
3.For the picture done triggered by sequence end, the ucode needs to
save the local memory and the status of some registers in the driver
does not need to be cleared.

Verify:
S5

SourceCode:
ucode:
0.4.51-ga799f55
I8dd73
I755e3
I0ac92
Ic96da
Id3122

Change-Id: I9e4839541843695840d7f0d307cb56d09477353c
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>