h264: CB2 decoder h264 avbc header size adjustment. [1/1]
PD#SWPL-121816

Problem:
The avbc header size on the decoder and
the avbc lib side are not aligned,
causing the map and flush on the avbc lib side
to exceed the size and cause an exception.

Solution:
The avbc header size is allocated in 1080/4k bins
and aligned with the avbc lib side.

Verify:
T5D

Change-Id: I72d9e4f8699315191ac35b50f086a9aba607b93b
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>