decoder: CB1 fix memory leak for hdr10p stream. [1/1]
PD#SWPL-122246

Problem:
A memory leak occurs because the address allocated to store the hdr10p
data is not released.

Solution:
Bind the hdr10p data to the callback function of codec_mm, and release
the hdr10p data when the yuv or avbc header buffer is finally released.

Verify:
ohm and s5

Change-Id: If1a3f2f7c342f9ad42c177c297ced4abe69508ee
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>