h265: CB2 fix the problem of long-term playback stuck. [1/1]
PD#SWPL-127049
PD#SWPL-132273

Problem:
1.The error stream causes the sps_max_dec_pic_buffering_minus1_0
information to become smaller, and the use count will be greater than
the pic num in run_ready. In this case, the playback will be stuck
without adjustment.
2.The error stream causes the calculated lcu information to be abnormal,
so the buffer size calculated through the lcu information will be
abnormally large, which will lead to memory allocation failure, and the
corresponding DECODER_FATAL_ERROR_NO_MEM bit of fatal_error will be set,
he decoder will not be able to run and the playback will be stuck.

Solution:
1.For h265 resolution change stream, the decoding driver will release
the previous buffer first, and then reallocate the new buffer, It is not
necessary to limit decoding by use count and pic num in run_ready.
2.By checking the size information of lcu, it returns directly without
continuing to decode when it is abnormal.

Verify:
S5.

Change-Id: I7b8bc69ce97e32faecb39acc95664f72f07540cb
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>