media_modules: CB2 H.264 fix interlace error stream display logic [1/1]
PD#SWPL-123839

Problem:
video stuck when display the error interlace stream, because the top field
and bottom field may be sent dividedly, so when the buffer is put back,
the vf_ref will be less than 0, such as -1 or -2.

Solution:
modify the error field display logic. the single field will not be
displayed until the next another picture is decoded. or the both fields
will not be displayed until the top and botton fields are all decoded.

Verify:
ohm

Change-Id: I389c5ae1229c60e6944bf1995251840935743855
Signed-off-by: xiaohang.cui <xiaohang.cui@amlogic.com>