h265: CB2 fix decoder error problems. [1/1]
PD#SWPL-126198

Problem:
The end_of_seq information is one frame before, and the decoder cannot
continue decoding due to the inaccurate HEVC_SHIFT_BYTE_COUNT value,
and the playback is stuck.

Solution:
For the s5 platform, shiftbyte needs to be reduced by 8bytes, and the
decoding driver recognizes that the end_of_seq information is one frame
head, so the ucode will continue to decode.

Verify:
S5.

Change-Id: I959a119a156e3c2fe333c236bb4b1fb29063e34b
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>