decoder: CB1 Crop stream afbc display abnormal. [1/1]
PD#SWPL-127648

Problem:
The avbc on the decoder side is compressed
according to the original size of 256x144.
At present, the width and height passed by the decoder
through vf is 176x144 after cropping, and the avbc decoding is abnormal.

Solution:
The decoder alloc buffers according to the original size 64 alignment,
and at the same time, vf passes the original width and height of afbc.

Verify:
ohm

Change-Id: Iff24ad727865f97d59340f32cafd57ad7c3f7925
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>