h265: CB2 fix decoder error problems for special stream. [1/1]
PD#SWPL-126522

Problem:
In the ucode, when the penultimate lcu is decoded, the pic_end_flag
variable will be set to 1. This will lead to when the last lcu is parsed,
since the pic_end_flag variable is 1, the hardware will not continue to
decode and one less lcu will be decoded.

Solution:
When the pic_end_flag variable is 1, check the column coordinates of the
lcu to see if all the lcu have been decoded. If not, it will continue to
decode.

Verify:
ohm

SourceCode:
ucode:
0.4.63-g3eba222
I95049
I696b7
I808ff
I00629
I92b70

Change-Id: Icf2d11db231159b01a93ea6720a356c60f25fd4a
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>