av1: CB2 dual-core playback stuck issue [1/1]
PD#SWPL-129361

Problem:
The file include show_existing_frame flag, back_done_mark is
incorrectly set to 0(vf_put function). As a result, it cannot
be sent for the second time, cause playback stuck.

Solution:
back_done_mark is set to 0 in new_pic, this operation
does not need to be performed in vf_put function.
Delete the back_done_mark operation set to 0 in vf_put
to resolve this problem

Verify:
S5

Change-Id: I36f1792f1d0dcdf1b371a65be4b2690c205e3541
Signed-off-by: Peng Yixin <yixin.peng@amlogic.com>