decoder: CB2 fix playback black screen problem caused by unscheduling. [1/1]
PD#SWPL-128742

Problem:
Because other processes occupy the cpu for too long, the vdec_thread
process will not be scheduled when the playback exits, which will cause
the sched_mask state occupied by the decoding instance to not be cleared,
so that the next playback decoder cannot be scheduled and playback will
be stuck.

Solution:
When the playback exits, if there is an abnormal decoding instance, the
corresponding sched_mask of the core is cleared by using the sched_mask
information of the abnormal instance.

Verify:
T5M

Change-Id: I1e581d937c2fefb94d58bdc4c65bda86a9cd3732
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>