media_modules: CB2 H.264 fix wrong poc sequence when stream loop back [1/1]
PD#SWPL-127464

Problem:
the poc of stream in the end is much greater than
the poc of stream in the start with dirty data.
and reorder is 2, when looping back, the output
poc sequence is not correct.

Solution:
When the difference of the POCs in looping back
is over 300, the frame of larger poc in the dpb
will be marked as displayed to not output.

Verify:
ohm

Change-Id: I49268b31a7442e582c709c79e58d744e25567e25
Signed-off-by: xiaohang.cui <xiaohang.cui@amlogic.com>