h264: CB2 Force interlacing when playing special streams [1/1]
PD#SWPL-136815

Problem:
The SEI information of the error flow is different from
the I/P flag in the SPS information, causing di to
start incorrectly.

Solution:
When the SEI information is an interlaced stream and di
is enabled, the interlaced stream output logic is
forced to be used.

Verify:
s4

Change-Id: I67fc884d64cc48b2247a18bee373f39df1b46206
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>