decoder: CB1 fix transfer process of userdata [1/1]
PD#SWPL-132588

Problem:
After the decoder driver decodes userdata, the work queue of userdata
is scheduled slowly, and the run is cleared when the upper-layer reads
the data. As a result, the upper-layer cannot obtain the userdata.

Solution:
decoder driver no longer uses the work queue to store userdata,
but fills cc data directly to ensure that the data is correct.

Verify:
ah212

Change-Id: I04025661ade47c6a8e5bc825b362bc1461df9137
Signed-off-by: rong.quan <rong.quan@amlogic.com>