v4ldec: CF1 dma buffer alloc delay for interlaced stream. [3/3]
PD#SWPL-119524

Problem:
Due to the post-positioning of di,
the vb2 buffer is directly thrown from the decoder to the upper layer.
Since each vb2 buffer can only be returned by the upper layer,
it can be thrown to the upper layer again. For interlaced streams,
two fields need to be output, so the second field cannot be output.

Solution:
The capture dma buffer is allocated in the decoder.
uvm only allocates an empty shell,
which is actually allocated by the decoder and replaces the dma buffer.

Verify:
ohm

Change-Id: I9b034a0cf505eb798595f2458c839b2398bcee80
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>