h265: CB2 h265 freezing playing with prime video. [1/1]
PD#SWPL-137723

Problem:
The frame division of prime video is different.
Some non-frame nals are sent to the decoder as a frame.
The decoder does not decode the frame and
does not report the timestamp,
causing the upper layer work to leak and not send es data,
causing it to get stuck.

Solution:
In this case, the decoder needs to report timestamp.

Verify:
ohm

Change-Id: I4b412929f403ce2c2a3b710440bab88d61d9512e
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>