avs2: CB2 fix palyback problems for some special streams. [1/1]
PD#SWPL-133980

Problem:
Since the player divides the sequence header and sequence endr into two
frames of data and sends them to the decoder.
1.The ucode does not save the width and height information of the stream
after parsing the sequence header information. In this way, when parsing
the frame header, it is detected that the width and height information
is abnormal and a decoding exception occurs.
2.The decoder continues to decode after processing the corresponding
sequence interrupt. Due to the decoding exception due to the frames size
information, so frames will be lost and the playback will not be smooth.

Solution:
1.When parsing the sequence header, you need to save the lmem information.
2.After processing the interrupts of sequence header and sequence end,
the size information of frames is used to determine whether decoding
needs to continue.

Verify:
S5.

SourceCode:
ucode:
0.4.95-g5a2b348
If6089
I4783e
I3aed8
I5c658
Id2ef4

Change-Id: I89423e8fe7531b138f990f0a5239e68bd9b620cb
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>