avs3: CB2 fix first frame output slow problem. [1/1]
PD#SWPL-148133

Problem:
The first decoded frame was discarded when it was put in the output
buffer, because it was detected that the first frame data was already in
the output buffer.This would cause the subsequent data that referenced
that frame to be discarded as well, resulting in the slow output of the
first frame.

Solution:
Optimize the output buffer checking mechanism to solve the problem of
the first frame data being discarded.

Verify:
S5.

Change-Id: Ibab27a6ba287d9c9d8c7e0934cba02366cefce99
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>