h264: CB2 Variable resolution optimization when seek [1/3]
PD#SWPL-146844

Problem:
Seek occurs during the resolution changing process,
resulting in the decoder resolution changing process not being completed.
The two parsing headers may be inconsistent or
the resolution event may not be reported successfully,
causing the upper layer and the decoder code stream information status
to be inconsistent and stuck.

Solution:
When changing resolution, if the resolution event is
not reported successfully,
the decoder will discard the resolution event;
if the resolution is reported successfully,
a double check will be performed when parsing the sequence header
for the second time.

Verify:
T7C

Change-Id: I66d571257a4253047a33f8b0b739d25d82366cb6
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>