v4l: CF2 Added duration reporting function [1/1]
PD#SWPL-152674

Problem:
When duration changes in the stream,
the upper layer cannot recognize it,
resulting in a delay in subsequent
playback.

Solution:
When the decoder resolves to the duration
transform in the stream, it notifies the
upper layer, so that the upper layer can
respond in time.

Verify:
T3X

Change-Id: I74e992f0d64f75d52d100ef382723c0ca3e31b93
Signed-off-by: kuan.hu <kuan.hu@amlogic.com>