h265: CB2 Seek scene frame loss and stuck problem [1/1]
PD#SWPL-159997
PD#SWPL-162777

Problem:
In the seek scene of h265 pair pic stream,
since the first frame is composed of the top and bottom fields combined with
the p field and the i field respectively,
the upper layer only sends down the i field, and the i field poc value is 9,
so the entire gop is discarded.
When decoding the next gop,
since the poc value of the first frame of the previous gop is 9,
frames with poc value less than 9 in the next gop cannot be output.

Solution:
When the IDR frame is decoded,
the POC value record of the first frame is updated.

Verify:
TXHD2

Change-Id: Iea43fed52fda02f2fe86c5779bf02f5a984e0a64
Signed-off-by: miaohong chen <miaohong.chen@amlogic.com>