h265: CB2 fix playback unsmooth problems for special stream. [1/1]
PD#SWPL-162604

Problem:
The abnormal decoding and frame dropping causing playback unsmooth are
due to the fact that the  value read by shiftbyte is 8 bytes larger than
actually used. This results in an  abnormal judgment of eos in head in
the decoding driver.

Solution:
Increase the shiftbyte value of eos in head in the decoding driver to
resolve the decoding abnormality caused by the inaccurate shiftbyte.

Verify:
SC2.

Change-Id: I21b791357ac42a81f1f4caea491e063e778ea273
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>