h264: CB1 fix the problem of stuck playback for special streams. [1/1]
PD#SWPL-168639

Problem:
The stream switches from progressive to interlaced, so that the double
write of the decoding driver switches from mmu to no-mmu. This causes a
decoding exception because the mmu related configuration is not cleared.

Solution:
When the double write of the decoding driver switches from mmu to no-mmu,
clear the mmu related configuration.

Verify:
S5

Change-Id: I8981a0d586de6b3135f8b49bc97b585af422d6c9
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>